# Doc/swagger.yaml
openapi: 3.0.3
info:
  title: Green Dii API
  description: |
    REST API for Green Dii Sales Workflow System
    ระบบจัดการ Sales Workflow — Quotation → Work Order → Hand Over → Purchase Request
    **Default password for all seeded users: `1234`**
  version: 1.0.0
  contact:
    name: Green Dii Support
servers:
  - url: http://localhost:4000/api
    description: Development
  - url: https://your-domain.com/api
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:         { type: string }
        username:   { type: string }
        fullName:   { type: string }
        initials:   { type: string }
        role:       { type: string, enum: [sales, sales2, sale_mgr, admin_mgr, project_mgr, director, procurement, factory] }
        lineUserId: { type: string }
        active:     { type: boolean }
        createdAt:  { type: string, format: date-time }
    Customer:
      type: object
      properties:
        id:            { type: string }
        name:          { type: string }
        contactPerson: { type: string }
        tel:           { type: string }
        email:         { type: string }
        address:       { type: string }
        taxId:         { type: string }
        type:          { type: string }
        active:        { type: boolean }
    Product:
      type: object
      properties:
        id:          { type: string }
        code:        { type: string }
        name:        { type: string }
        category:    { type: string }
        unit:        { type: string }
        price:       { type: number }
        cost:        { type: number }
        description: { type: string }
        active:      { type: boolean }
...
    HandOverJob:
      type: object
      properties:
        id:          { type: string }
        hoNo:        { type: string }
        project:     { type: string }
        status:      { type: string, enum: [draft, pending, approved] }
        serviceDate: { type: string, format: date }
...
    Error:
      type: object
      properties:
        message: { type: string }
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, example: somchai }
                password: { type: string, example: "1234" }
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:  { $ref: '#/components/schemas/User' }
        "401": { description: Invalid credentials, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: { type: string }
                newPassword: { type: string }
      responses:
        "200": { description: Password changed }
        "400": { description: Wrong old password }

  # ── USERS ────────────────────────────────────────────────────────────────────
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - in: query
          name: active
          schema: { type: boolean }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200": { description: Customer list, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Customer' } } } } }
    post:
      tags: [Customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Customer' }
      responses:
        "201": { description: Customer created }
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Customer }
    put:
      tags: [Customers]
      summary: Update customer
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Customer' }
      responses:
        "200": { description: Updated }
    delete:
      tags: [Customers]
      summary: Deactivate customer
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Deactivated }

  # ── PRODUCTS ──────────────────────────────────────────────────────────────────
  /products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - in: query
          name: active
          schema: { type: boolean }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200": { description: Product list }
    post:
      tags: [Products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Product' }
      responses:
        "201": { description: Created }
  /products/{id}:
    put:
      tags: [Products]
      summary: Update product
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Product' }
      responses:
        "200": { description: Updated }
    delete:
      tags: [Products]
      summary: Deactivate product
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Deactivated }

  # ── UNITS ─────────────────────────────────────────────────────────────────────
...
  /handovers:
    get:
      tags: [HandOver]
      summary: List hand over jobs
      responses:
        "200": { description: List }
    post:
      tags: [HandOver]
      summary: Create hand over job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HandOverJob' }
      responses:
        "201": { description: Created }
  /handovers/{id}:
    put:
      tags: [HandOver]
      summary: Update hand over job
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        "200": { description: Updated }

  # ── PURCHASE REQUESTS ─────────────────────────────────────────────────────────
...